Refactor navbar filters component and service
- Updated navbar-filters.component.html for improved readability and structure.
- Refactored navbar-filters.component.ts to utilize signals and remove unused imports.
- Simplified navbar.component.html by removing unnecessary filter display logic.
- Enhanced navbar.component.ts to manage filter visibility and state more effectively.
- Introduced filters.service.ts to centralize filter logic and state management.
- Cleaned up navbar.service.ts by removing redundant filter logic and integrating with the new filters service.
