+++ src/app/pages/home-page/home-page.component.css
@@ -1,8 +1,27 @@
+.filters-floating-panel {
+  position: fixed;
+  top: calc(50% + 25px);
+  transform: translateY(-50%);
+  left: 20px;
+  width: 420px;
+  background: transparent;
+  border: none;
+  box-shadow: 0 12px 32px rgba(0, 0, 0, 0.35);
+  z-index: 1000;
+  overflow: auto;
+  transition: left 0.25s cubic-bezier(0.4, 0, 0.2, 1), opacity 0.25s ease,
+    transform 0.25s ease;
+}
+
+.filters-floating-panel.shifted {
+  left: 340px;
+}
+
 .home-page-container {
   height: 100vh;
   width: 100vw;
-  background: #f3f3f3;
   overflow: hidden;
+  background: transparent !important;
 }
 
 mat-drawer-content {
@@ -10,6 +29,7 @@ mat-drawer-content {
   width: 100%;
   display: flex;
   flex-direction: column;
+  background: transparent !important;
 }
 
 app-map {
@@ -21,9 +41,9 @@ app-map {
 .details-panel {
   width: 49vw;
   min-width: 360px;
-  background: transparent;
-  box-shadow: none;
-  border-left: none;
+  background: transparent !important;
+  box-shadow: none !important;
+  border-left: none !important;
   min-height: 0;
   padding: 20px 0;
   display: flex;
@@ -33,6 +53,11 @@ app-map {
   gap: 0;
   height: 95vh;
   overflow: hidden;
+}
+
+.details-panel mat-drawer-inner-container {
+  background: transparent !important;
+  box-shadow: none !important;
 }
 
 .building-card {
@@ -234,8 +259,12 @@ mat-grid-list {
   margin-right: 10px;
 }
 
-.properties-section h3 {
-  margin-top: 0;
+.properties-header {
+  display: flex;
+  gap: 2px;
+  align-items: center;
+  margin: 0;
+  padding: 4px 8px;
 }
 
 .property-item {
@@ -499,12 +528,13 @@ mat-grid-list {
 
 .gallery-modal-overlay {
   position: fixed;
-  z-index: 2000;
   inset: 0;
   background: rgba(0, 0, 0, 0.85);
   display: flex;
   align-items: center;
   justify-content: center;
+  z-index: 1200;
+  pointer-events: auto;
 }
 
 .gallery-modal-content {
@@ -517,6 +547,7 @@ mat-grid-list {
   display: flex;
   flex-direction: column;
   align-items: center;
+  padding: 16px;
 }
 
 .gallery-close-btn {
@@ -722,113 +753,6 @@ mat-grid-list.mat-grid-list > div {
   min-height: 0 !important;
 }
 
-@media (max-width: 1100px) {
-  .property-card {
-    width: clamp(250px, 48vw, 430px);
-    min-width: 180px;
-    max-width: 98vw;
-  }
-}
-
-@media (max-width: 800px) {
-  .property-card {
-    width: clamp(200px, 82vw, 430px);
-    min-width: 140px;
-    max-width: 99vw;
-    border-radius: 16px;
-    margin: 12px auto;
-  }
-  .property-image-container {
-    border-radius: 16px 16px 0 0;
-    min-height: 100px;
-  }
-}
-
-@media (max-width: 600px) {
-  .property-actions {
-    min-height: 38px;
-    padding: 0 6px;
-  }
-  .property-action,
-  .property-action-more {
-    height: 38px;
-    min-width: 38px;
-    font-size: 1.1rem;
-  }
-  .property-action-more {
-    padding: 0 8px;
-    font-size: 0.95rem;
-    border-radius: 13px;
-  }
-  .property-action mat-icon {
-    font-size: 22px;
-  }
-  .property-card {
-    width: clamp(170px, 90vw, 330px);
-    min-width: 110px;
-    margin: 8px auto;
-    border-radius: 12px;
-  }
-  .property-image-container,
-  .property-image {
-    border-radius: 12px 12px 0 0;
-    min-height: 80px;
-  }
-  .property-title-overlay {
-    padding: 8px 8px 4px 10px;
-    font-size: 0.95rem;
-    border-radius: 0 0 8px 8px;
-  }
-
-  .filters-fab {
-    right: 88px;
-    bottom: 24px;
-  }
-
-  .comparative-fab {
-    right: 24px;
-    bottom: 24px;
-  }
-}
-
-@media (max-width: 400px) {
-  .property-card {
-    width: 98vw;
-    min-width: 80px;
-  }
-  .property-image-container,
-  .property-image {
-    min-height: 50px;
-  }
-
-  .filters-fab {
-    right: 80px;
-    bottom: 20px;
-  }
-
-  .comparative-fab {
-    right: 20px;
-    bottom: 20px;
-  }
-}
-
-@media (max-width: 380px) {
-  .property-card {
-    width: 98vw;
-    min-width: 80px;
-  }
-  .property-image-container,
-  .property-image {
-    min-height: 50px;
-  }
-}
-
-.property-action-more:hover {
-  background: linear-gradient(
-    90deg,
-    rgba(80, 120, 210, 0.31) 70%,
-    rgba(120, 130, 220, 0.16) 100%
-  ) !important;
-  box-shadow: 0 2px 18px rgba(120, 130, 210, 0.17);
-  transform: translateY(-2px) scale(1.04);
+.mat-drawer-backdrop {
+  display: none !important;
 }
