+++ src/app/shared/components/navbar/navbar.component.html
@@ -93,11 +93,3 @@
     }
   </div>
 </mat-toolbar>
-
-
-@if(showFilters() && isHomePage() && !sidebarOpened() && !propertyDetailsOpened() && !buildingDetailsOpened()) {
-  <div class="navbar-filters">
-    <app-navbar-filters></app-navbar-filters>
-  </div>
-}
-
